add question deletion test

+++ src/domain/forum/application/use-cases/delete-question.ts
@@ -2,6 +2,7 @@ import { IQuestionRepository } from "../repositories/questions-repository"
 
 interface DeleteQuestionUseCaseRequest {
   questionId: string
+  authorId:string
 }
 
 interface DeleteQuestionUseCaseResponse {}
@@ -10,10 +11,14 @@ export class DeleteQuestionUseCase {
   constructor(private questionsRepository: IQuestionRepository) {}
   async execute({
     questionId,
+    authorId
   }: DeleteQuestionUseCaseRequest): Promise<DeleteQuestionUseCaseResponse> {
     const question = await this.questionsRepository.findById(questionId)
     if (!question) {
       throw new Error("Question not found")
+    }
+    if (question.authorId.toString() !== authorId) {
+      throw new Error("Unauthorized")
     }
     await this.questionsRepository.delete(question)
     return {}
